+++ apps/game_services_web/lib/game_services_web/controllers/registration_controller.ex
@@ -1,6 +1,7 @@
 defmodule GameServicesWeb.RegistrationController do
   @moduledoc false
   use GameServicesWeb, :controller
+  require Logger
 
   def index(conn, _params) do
     render(conn, "index.html")
@@ -12,17 +13,22 @@ defmodule GameServicesWeb.RegistrationController do
            Map.get(user_params, "email"),
            Map.get(user_params, "password")
          ) do
-      {:ok, _user} ->
+      {:ok, user} ->
+        Logger.info(fn -> "User Registration Success" end, user_id: user.id)
         redirect(conn, to: Routes.page_path(conn, :index))
 
       {:error, :create_credential, changeset, _other} ->
         err = Enum.map(changeset.errors, fn x -> get_errors_text(x) end)
+        Logger.metadata(register_error: err)
+        Logger.warn(fn -> "User Register Error" end)
 
         conn
         |> put_flash(:register_err, "Invalid Register Credentials - #{Enum.join(err, " , ")} -- ")
         |> render("index.html")
 
       _other ->
+        Logger.error(fn -> "Unknown Error" end)
+
         conn
         |> put_flash(:register_err, "unknown error")
         |> render("index.html")