+++ apps/tcp_server/lib/tcp_server/workflows/cmsg_authenticate_challenge.ex
@@ -1,0 +1,2 @@
+defmodule TcpServer.Workflows.CmsgAuthenticateChallenge do
+end