+++ apps/tcp_server/test/tcp_server_test.exs
@@ -6,15 +6,19 @@ defmodule TcpServerTest do
   @tag :wip
   test "sometthing" do
     {:ok, conn} = :gen_tcp.connect('localhost', 8005, [:binary, active: false])
-    id = 47
-    m = <<"hello">>
-    sz = byte_size(m)
+    id = 1
+    username = <<"tester123">>
+    username_len = byte_size(username)
     id_sz = 8 * 8
     sz_sz = 32
-    packet =  <<id::size(id_sz), sz::little-size(sz_sz)>> <> m
+
+    random_id = :rand.uniform(999_999)
+
+    message = <<username_len::size(8)>> <> username <> <<random_id::size(32)>>
+    sz = byte_size(message)
+
+    packet = <<id::size(id_sz), sz::little-size(sz_sz)>> <> message
     :gen_tcp.send(conn, packet)
-
-
 
     case :gen_tcp.recv(conn, 0, 5000) do
       {:ok, _data} ->