+++ apps/user_manager/lib/user_manager.ex
@@ -6,9 +6,13 @@ defmodule UserManager do
   @doc """
     Register a New User Account
   """
+  require Logger
+
   @spec register_new_user(String.t(), String.t(), String.t()) ::
           {:ok, %GameServices.Account.User{}} | {:error, atom(), any(), any()}
   def register_new_user(name, email, password) do
+    Logger.info(fn -> "Register User Request" end, name: name, email: email, password: password)
+
     GenServer.call(
       ClusterManager.get_registration_worker(),
       {:register_user, name, email, password}