+++ config/dev.exs
@@ -1,7 +1,10 @@
 use Mix.Config
 
 # Do not include metadata nor timestamps in development logs
-config :logger, :console, format: "[$level] $message\n"
+config :logger, :console,
+  format: {LogFormatter, :format},
+  level: :info,
+  metadata: :all
 
 # Initialize plugs at runtime for faster development compilation
 config :phoenix, :plug_init_mode, :runtime