+++ config/test.exs
@@ -1,4 +1,7 @@
 use Mix.Config
 
 # Print only warnings and errors during test
-config :logger, level: :warn
+config :logger,
+  level: :error,
+  format: {LogFormatter, :format},
+  metadata: :all